pyproject.toml
[tool.poetry]
name = "forward-summarizer-bot"
version = "0.1.0"
description = "Telegram bot with LLM summarization and categorization"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "forward_summarizer_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.7"
anthropic = "^0.18.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "bot:start"